Changes to move to fictional REST endpoint.
