--- e2e_samples/parking_sensors/adf/dataset/Ds_Http_Parking_Bay.json
@@ -1,21 +1,0 @@
-{
-    "name": "Ds_Http_Parking_Bay",
-    "properties": {
-        "linkedServiceName": {
-            "referenceName": "Ls_Http_Parking_Bay_01",
-            "type": "LinkedServiceReference"
-        },
-        "annotations": [],
-        "type": "Json",
-        "typeProperties": {
-            "location": {
-                "type": "HttpServerLocation",
-                "relativeUrl": {
-                    "value": "on-street-parking-bays/exports/json",
-                    "type": "Expression"
-                }
-            }
-        },
-        "schema": {}
-    }
-}
