--- e2e_samples/parking_sensors/adf/dataset/Ds_Http_Parking_Bay_Sensors.json
@@ -1,18 +1,0 @@
-{
-    "name": "Ds_Http_Parking_Bay_Sensors",
-    "properties": {
-        "linkedServiceName": {
-            "referenceName": "Ls_Http_Parking_Bay_01",
-            "type": "LinkedServiceReference"
-        },
-        "annotations": [],
-        "type": "Json",
-        "typeProperties": {
-            "location": {
-                "type": "HttpServerLocation",
-                "relativeUrl": "on-street-parking-bay-sensors/exports/json"
-            }
-        },
-        "schema": {}
-    }
-}
