+++ e2e_samples/parking_sensors/adf/dataset/Ds_Http_Parking_Locations.json
@@ -1,0 +1,21 @@
+{
+    "name": "Ds_Http_Parking_Locations",
+    "properties": {
+        "linkedServiceName": {
+            "referenceName": "Ls_Http_DataSimulator",
+            "type": "LinkedServiceReference"
+        },
+        "annotations": [],
+        "type": "Json",
+        "typeProperties": {
+            "location": {
+                "type": "HttpServerLocation",
+                "relativeUrl": {
+                    "value": "/locations",
+                    "type": "Expression"
+                }
+            }
+        },
+        "schema": {}
+    }
+}
