+++ e2e_samples/parking_sensors/adf/dataset/Ds_Http_Parking_Sensors.json
@@ -1,0 +1,18 @@
+{
+    "name": "Ds_Http_Parking_Sensors",
+    "properties": {
+        "linkedServiceName": {
+            "referenceName": "Ls_Http_DataSimulator",
+            "type": "LinkedServiceReference"
+        },
+        "annotations": [],
+        "type": "Json",
+        "typeProperties": {
+            "location": {
+                "type": "HttpServerLocation",
+                "relativeUrl": "/sensors"
+            }
+        },
+        "schema": {}
+    }
+}
