--- e2e_samples/parking_sensors/adf/dataset/Ds_REST_MelbParkingData.json
@@ -1,24 +1,0 @@
-{
-	"name": "Ds_REST_MelbParkingData",
-	"properties": {
-		"linkedServiceName": {
-			"referenceName": "Ls_Rest_MelParkSensors_01",
-			"type": "LinkedServiceReference"
-		},
-		"parameters": {
-			"relativeurl": {
-				"type": "String"
-			}
-		},
-		"annotations": [],
-		"type": "RestResource",
-		"typeProperties": {
-			"relativeUrl": {
-				"value": "@dataset().relativeurl",
-				"type": "Expression"
-			},
-			"requestMethod": "GET"
-		}
-	},
-	"type": "Microsoft.DataFactory/factories/datasets"
-}
