+++ e2e_samples/parking_sensors/adf/dataset/Ds_REST_ParkingData.json
@@ -1,0 +1,24 @@
+{
+	"name": "Ds_REST_ParkingData",
+	"properties": {
+		"linkedServiceName": {
+			"referenceName": "Ls_Rest_ParkSensors_01",
+			"type": "LinkedServiceReference"
+		},
+		"parameters": {
+			"relativeurl": {
+				"type": "String"
+			}
+		},
+		"annotations": [],
+		"type": "RestResource",
+		"typeProperties": {
+			"relativeUrl": {
+				"value": "@dataset().relativeurl",
+				"type": "Expression"
+			},
+			"requestMethod": "GET"
+		}
+	},
+	"type": "Microsoft.DataFactory/factories/datasets"
+}
