+++ e2e_samples/parking_sensors/adf/linkedService/Ls_Http_DataSimulator.json.template
@@ -1,0 +1,12 @@
+{
+    "name": "Ls_Http_DataSimulator",
+    "properties": {
+        "annotations": [],
+        "type": "HttpServer",
+        "typeProperties": {
+            "url": "https://<apiBaseUrl>",
+            "enableServerCertificateValidation": true,
+            "authenticationType": "Anonymous"
+        }
+    }
+}
