--- e2e_samples/parking_sensors/adf/linkedService/Ls_Http_Parking_Bay_01.json
@@ -1,12 +1,0 @@
-{
-    "name": "Ls_Http_Parking_Bay_01",
-    "properties": {
-        "annotations": [],
-        "type": "HttpServer",
-        "typeProperties": {
-            "url": "https://data.melbourne.vic.gov.au/api/explore/v2.1/catalog/datasets/",
-            "enableServerCertificateValidation": true,
-            "authenticationType": "Anonymous"
-        }
-    }
-}
