--- e2e_samples/parking_sensors/adf/linkedService/Ls_Rest_MelParkSensors_01.json
@@ -1,13 +1,0 @@
-{
-	"name": "Ls_Rest_MelParkSensors_01",
-	"type": "Microsoft.DataFactory/factories/linkedservices",
-	"properties": {
-		"annotations": [],
-		"type": "RestService",
-		"typeProperties": {
-			"url": "https://data.melbourne.vic.gov.au/resource/",
-			"enableServerCertificateValidation": true,
-			"authenticationType": "Anonymous"
-		}
-	}
-}
