+++ e2e_samples/parking_sensors/adf/linkedService/Ls_Rest_ParkSensors_01.json.template
@@ -1,0 +1,13 @@
+{
+	"name": "Ls_Rest_ParkSensors_01",
+	"type": "Microsoft.DataFactory/factories/linkedservices",
+	"properties": {
+		"annotations": [],
+		"type": "RestService",
+		"typeProperties": {
+			"url": "https://<apiBaseUrl>",
+			"enableServerCertificateValidation": true,
+			"authenticationType": "Anonymous"
+		}
+	}
+}
