--- e2e_samples/parking_sensors/adf/pipeline/P_Ingest_MelbParkingData.json
@@ -1,275 +1,0 @@
-{
-	"name": "P_Ingest_MelbParkingData",
-	"properties": {
-		"description": "Hello from NDC Sydney!!!",
-		"activities": [
-			{
-				"name": "StandardizeData",
-				"description": "",
-				"type": "DatabricksNotebook",
-				"dependsOn": [
-					{
-						"activity": "DownloadSensorData",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					},
-					{
-						"activity": "DownloadBayData",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					}
-				],
-				"policy": {
-					"timeout": "7.00:00:00",
-					"retry": 0,
-					"retryIntervalInSeconds": 30,
-					"secureOutput": false,
-					"secureInput": false
-				},
-				"userProperties": [],
-				"typeProperties": {
-					"notebookPath": "/notebooks/02_standardize",
-					"baseParameters": {
-						"infilefolder": {
-							"value": "@variables('infilefolder')",
-							"type": "Expression"
-						},
-						"loadid": {
-							"value": "@pipeline().RunId",
-							"type": "Expression"
-						}
-					},
-					"libraries": [
-						{
-							"egg": "dbfs:/ddo_transform-localdev-py2.py3-none-any.whl"
-						},
-						{
-							"pypi": {
-								"package": "applicationinsights"
-							}
-						}
-					]
-				},
-				"linkedServiceName": {
-					"referenceName": "Ls_AzureDatabricks_01",
-					"type": "LinkedServiceReference"
-				}
-			},
-			{
-				"name": "Set infilefolder",
-				"type": "SetVariable",
-				"dependsOn": [],
-				"userProperties": [],
-				"typeProperties": {
-					"variableName": "infilefolder",
-					"value": {
-						"value": "@utcnow('yyyy_MM_dd_hh_mm_ss')",
-						"type": "Expression"
-					}
-				}
-			},
-			{
-				"name": "DownloadSensorData",
-				"type": "Copy",
-				"dependsOn": [
-					{
-						"activity": "Set infilefolder",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					}
-				],
-				"policy": {
-					"timeout": "7.00:00:00",
-					"retry": 0,
-					"retryIntervalInSeconds": 30,
-					"secureOutput": false,
-					"secureInput": false
-				},
-				"userProperties": [],
-				"typeProperties": {
-					"source": {
-                        "type": "JsonSource",
-                        "storeSettings": {
-                            "type": "HttpReadSettings",
-                            "requestMethod": "GET"
-                        },
-                        "formatSettings": {
-                            "type": "JsonReadSettings"
-                        }
-                    },
-					"sink": {
-						"type": "AzureBlobFSSink"
-					},
-					"enableStaging": false
-				},
-				"inputs": [
-					{
-						"referenceName": "Ds_Http_Parking_Bay",
-						"type": "DatasetReference"
-					}
-				],
-				"outputs": [
-					{
-						"referenceName": "Ds_AdlsGen2_MelbParkingData",
-						"type": "DatasetReference",
-						"parameters": {
-							"infilefolder": {
-								"value": "@variables('infilefolder')",
-								"type": "Expression"
-							},
-							"infilename": "MelbParkingSensorData.json",
-							"container": "datalake/data/lnd"
-						}
-					}
-				]
-			},
-			{
-				"name": "DownloadBayData",
-				"type": "Copy",
-				"dependsOn": [
-					{
-						"activity": "Set infilefolder",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					}
-				],
-				"policy": {
-					"timeout": "7.00:00:00",
-					"retry": 0,
-					"retryIntervalInSeconds": 30,
-					"secureOutput": false,
-					"secureInput": false
-				},
-				"userProperties": [],
-				"typeProperties": {
-                    "source": {
-                        "type": "JsonSource",
-                        "storeSettings": {
-                            "type": "HttpReadSettings",
-                            "requestMethod": "GET"
-                        },
-                        "formatSettings": {
-                            "type": "JsonReadSettings"
-                        }
-                    },
-					"sink": {
-						"type": "AzureBlobFSSink"
-					},
-					"enableStaging": false
-				},
-				"inputs": [
-					{
-						"referenceName": "Ds_Http_Parking_Bay_Sensors",
-						"type": "DatasetReference"
-					}
-				],
-				"outputs": [
-					{
-						"referenceName": "Ds_AdlsGen2_MelbParkingData",
-						"type": "DatasetReference",
-						"parameters": {
-							"infilefolder": {
-								"value": "@variables('infilefolder')",
-								"type": "Expression"
-							},
-							"infilename": "MelbParkingBayData.json",
-							"container": "datalake/data/lnd"
-						}
-					}
-				]
-			},
-			{
-				"name": "TransformData",
-				"type": "DatabricksNotebook",
-				"dependsOn": [
-					{
-						"activity": "StandardizeData",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					}
-				],
-				"policy": {
-					"timeout": "7.00:00:00",
-					"retry": 0,
-					"retryIntervalInSeconds": 30,
-					"secureOutput": false,
-					"secureInput": false
-				},
-				"userProperties": [],
-				"typeProperties": {
-					"notebookPath": "/notebooks/03_transform",
-					"baseParameters": {
-						"loadid": {
-							"value": "@pipeline().RunId",
-							"type": "Expression"
-						}
-					},
-					"libraries": [
-						{
-							"egg": "dbfs:/ddo_transform-localdev-py2.py3-none-any.whl"
-						},
-						{
-							"pypi": {
-								"package": "applicationinsights"
-							}
-						}
-					]
-				},
-				"linkedServiceName": {
-					"referenceName": "Ls_AzureDatabricks_01",
-					"type": "LinkedServiceReference"
-				}
-			},
-			{
-				"name": "Load SQLDW",
-				"type": "SqlServerStoredProcedure",
-				"dependsOn": [
-					{
-						"activity": "TransformData",
-						"dependencyConditions": [
-							"Succeeded"
-						]
-					}
-				],
-				"policy": {
-					"timeout": "7.00:00:00",
-					"retry": 0,
-					"retryIntervalInSeconds": 30,
-					"secureOutput": false,
-					"secureInput": false
-				},
-				"userProperties": [],
-				"typeProperties": {
-					"storedProcedureName": "[dbo].[load_dw]",
-					"storedProcedureParameters": {
-						"load_id": {
-							"value": {
-								"value": "@pipeline().RunId",
-								"type": "Expression"
-							},
-							"type": "String"
-						}
-					}
-				},
-				"linkedServiceName": {
-					"referenceName": "Ls_AzureSQLDW_01",
-					"type": "LinkedServiceReference"
-				}
-			}
-		],
-		"variables": {
-			"infilefolder": {
-				"type": "String",
-				"defaultValue": "lnd"
-			}
-		},
-		"annotations": [],
-		"lastPublishTime": "2021-08-03T05:18:18Z"
-	},
-	"type": "Microsoft.DataFactory/factories/pipelines"
-}
