+++ e2e_samples/parking_sensors/adf/trigger/T_Sched.json
@@ -6,7 +6,7 @@
 		"pipelines": [
 			{
 				"pipelineReference": {
-					"referenceName": "P_Ingest_MelbParkingData",
+					"referenceName": "P_Ingest_ParkingData",
 					"type": "PipelineReference"
 				}
 			}
