+++ e2e_samples/parking_sensors/databricks/notebooks/02_standardize.py
@@ -22,8 +22,8 @@ import datetime
 load_id = loadid
 loaded_on = datetime.datetime.now()
 base_path = os.path.join('dbfs:/mnt/datalake/data/lnd/', infilefolder)
-parkingbay_filepath = os.path.join(base_path, "MelbParkingBayData.json")
-sensors_filepath = os.path.join(base_path, "MelbParkingSensorData.json")
+parkingbay_filepath = os.path.join(base_path, "ParkingLocationData.json")
+sensors_filepath = os.path.join(base_path, "ParkingSensorData.json")
 
 
 # COMMAND ----------
@@ -31,18 +31,18 @@ sensors_filepath = os.path.join(base_path, "MelbParkingSensorData.json")
 import ddo_transform.standardize as s
 
 # Retrieve schema
-parkingbay_schema = s.get_schema("in_parkingbay_schema")
+parkingbay_schema = s.get_schema("in_parkinglocation_schema")
 sensordata_schema = s.get_schema("in_sensordata_schema")
 
 # Read data
 parkingbay_sdf = spark.read\
   .schema(parkingbay_schema)\
-  .option("badRecordsPath", os.path.join(base_path, "__corrupt", "MelbParkingBayData"))\
+  .option("badRecordsPath", os.path.join(base_path, "__corrupt", "ParkingLocationData"))\
   .option("multiLine", True)\
   .json(parkingbay_filepath)
 sensordata_sdf = spark.read\
   .schema(sensordata_schema)\
-  .option("badRecordsPath", os.path.join(base_path, "__corrupt", "MelbParkingSensorData"))\
+  .option("badRecordsPath", os.path.join(base_path, "__corrupt", "ParkingSensorData"))\
   .option("multiLine", True)\
   .json(sensors_filepath)
