+++ e2e_samples/parking_sensors/devops/templates/jobs/deploy-adf-job.yml
@@ -36,7 +36,7 @@ jobs:
             location: '$(azureLocation)'
             csmFile: '$(Pipeline.Workspace)/ciartifacts/adf_arm_template/ARMTemplateForFactory.json'
             csmParametersFile: '$(Pipeline.Workspace)/ciartifacts/adf_arm_template/ARMTemplateParametersForFactory.json'
-            overrideParameters: -factoryName "$(adfName)" -P_Ingest_MelbParkingData_properties_0_typeProperties_notebookPath "$(databricksNotebookPath)/02_standardize" -P_Ingest_MelbParkingData_properties_0_typeProperties_0_egg "$(databricksDbfsLibPath)/$(packageWheelName)" -P_Ingest_MelbParkingData_properties_4_typeProperties_0_egg "$(databricksDbfsLibPath)/$(packageWheelName)" -Ls_AdlsGen2_01_properties_typeProperties_url "https://$(datalakeAccountName).dfs.core.windows.net/" -Ls_KeyVault_01_properties_typeProperties_baseUrl "$(kvUrl)" -Ls_Rest_MelParkSensors_01_properties_typeProperties_url "$(apiBaseUrl)" -P_Ingest_MelbParkingData_properties_4_typeProperties_notebookPath "$(databricksNotebookPath)/03_transform" -Ls_AzureDatabricks_01_properties_typeProperties_domain "$(databricksDomain)" -Ls_AzureDatabricks_01_properties_typeProperties_workspaceResourceId "$(databricksWorkspaceResourceId)"
+            overrideParameters: -factoryName "$(adfName)" -P_Ingest_ParkingData_properties_0_typeProperties_notebookPath "$(databricksNotebookPath)/02_standardize" -P_Ingest_ParkingData_properties_0_typeProperties_0_egg "$(databricksDbfsLibPath)/$(packageWheelName)" -P_Ingest_ParkingData_properties_4_typeProperties_0_egg "$(databricksDbfsLibPath)/$(packageWheelName)" -Ls_AdlsGen2_01_properties_typeProperties_url "https://$(datalakeAccountName).dfs.core.windows.net/" -Ls_KeyVault_01_properties_typeProperties_baseUrl "$(kvUrl)" -Ls_Rest_ParkSensors_01_properties_typeProperties_url "$(apiBaseUrl)" -P_Ingest_ParkingData_properties_4_typeProperties_notebookPath "$(databricksNotebookPath)/03_transform" -Ls_AzureDatabricks_01_properties_typeProperties_domain "$(databricksDomain)" -Ls_AzureDatabricks_01_properties_typeProperties_workspaceResourceId "$(databricksWorkspaceResourceId)"
 
 
         # https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/deploy/azure-powershell
